reorganize and reuse code
