+++ backend/src/v1/routes/dashboard/announcement-metrics-routes.spec.ts
@@ -4,7 +4,7 @@ import router from './announcement-metrics-routes';
 
 let app: Application;
 const getAnnouncementMetricsMock = jest.fn();
-jest.mock('../../services/dashboard-metrics-service', () => ({
+jest.mock('../../services/announcements-service', () => ({
   getAnnouncementMetrics: (...args) => getAnnouncementMetricsMock(...args),
 }));
 describe('announcement-metrics-routes', () => {
