+++ backend/src/v1/services/announcements-service.spec.ts
@@ -8,7 +8,7 @@ import { UserInputError } from '../types/errors';
 import * as AnnouncementService from './announcements-service';
 import { utils } from './utils-service';
 import { LocalDateTime, ZonedDateTime, ZoneId } from '@js-joda/core';
-import { updateAnnouncement } from './announcements-service';
+import { getAnnouncementMetrics, updateAnnouncement } from './announcements-service';
 
 const mockFindMany = jest.fn().mockResolvedValue([
   {
@@ -46,6 +46,10 @@ jest.mock('../prisma/prisma-client', () => ({
       updateMany: (...args) => mockUpdateMany(...args),
       create: (...args) => mockCreateAnnouncement(...args),
       findUniqueOrThrow: (...args) => mockFindUniqueOrThrow(...args),
+      groupBy: jest.fn().mockResolvedValueOnce([
+        { status: 'PUBLISHED', _count: 1 },
+        { status: 'DRAFT', _count: 2 },
+      ]),
     },
     announcement_history: {
       create: (...args) => mockHistoryCreate(...args),
@@ -920,6 +924,19 @@ describe('AnnouncementsService', () => {
       expect(mockFindUniqueOrThrow).toHaveBeenCalledWith({
         where: { announcement_id: '1' },
         include: { announcement_resource: true },
+      });
+    });
+  });
+
+  describe('getAnnouncementMetrics', () => {
+    it('should return the announcement metrics', async () => {
+      // Act
+      const result = await getAnnouncementMetrics();
+
+      // Assert
+      expect(result).toEqual({
+        published: { count: 1 },
+        draft: { count: 2 },
       });
     });
   });
